backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==2.15.8
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.2
beautifulsoup4==4.12.2
billiard==4.1.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.7
cssselect==1.2.0
defusedxml==0.7.1
dill==0.3.7
distro==1.8.0
Django==4.2.3
django-allauth==0.58.2
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-timezone-field==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
filelock==3.13.1
frozenlist==1.4.0
google-ads==21.3.0
google-api-core==2.11.1
google-api-python-client==2.36.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kombu==5.3.1
lazy-object-proxy==1.9.0
lxml==4.9.3
Markdown==3.4.3
mccabe==0.7.0
multidict==6.0.4
oauthlib==3.2.2
openai==1.3.5
packaging==23.1
parsel==1.8.1
platformdirs==4.0.0
prompt-toolkit==3.0.39
Protego==0.3.0
proto-plus==1.22.3
protobuf==4.24.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycountry==22.3.5
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDispatcher==2.0.7
PyJWT==2.8.0
pylint==2.17.4
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
redis==5.0.0
referencing==0.30.0
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
Scrapy==2.11.0
scrapy-splash==0.9.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tldextract==5.1.1
tomlkit==0.12.3
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
w3lib==2.1.2
wcwidth==0.2.6
wrapt==1.16.0
yarl==1.9.2
zope.interface==6.1
